Make models a package instead of being a single (huge) module.
