+++ xonstat/views/search.py
@@ -1,16 +1,9 @@
-import datetime
 import logging
-import pyramid.httpexceptions
-import re
-import time
-from pyramid.response import Response
-from sqlalchemy import desc
-from sqlalchemy.orm.exc import MultipleResultsFound, NoResultFound
+
 from sqlalchemy import func
-from xonstat.models import *
-from xonstat.util import strip_colors, qfont_decode
-from xonstat.util import page_url, html_colors
-from webhelpers.paginate import Page, PageURL
+from webhelpers.paginate import Page
+from xonstat.models import DBSession, Server, Map, Game, PlayerGameStat, Player
+from xonstat.util import page_url
 
 log = logging.getLogger(__name__)
